fix(ui): fixed text when displaying a flashcard.
Fixed text displaying when a user is in learning mode. Before, if question or answer is long the earlier version didnt display the entire text. Now with the new implementation the text is breaking when it reaches the end of card so that the rest of the text follows on next row. made this change to make a better user experience.