build(build): Setting up REST-API architecture.
added folder server with additional subfolders (controller and service). in the subfolders we added needed files for later issues regarding REST-API setup. we needed to do this change so that the project can follow the REST-API standard, and these folders are needed for that. we also added a APIClient.java inside fxui/ui. This will later be needed for further development with REST.

Co-authored: @marennod
Closed #95