feat(ui): creating sign in feature.
Created sign up page and controller to ensure that new users can be created securely. This feature will ensure the user confirms the password and will only create a new user if the username is not already taken. Also implemented necessary updates to controllers and login system for this feature to work properly.

Co-authored: @ailinat
Closes #104