fix(ui): fix error with back button and sign up button
The back button on the FlashcardLearning page did not load the main page, and the sign up page was not loading. This was fixed by changing the button references in the controllers and updating the path to the page we want to load in the event handling methods. This fix is necessary for making the app functional and ensuring a good user experience.

#108

Co-authored-by: Maren Nodenes-Fimland <marennod@stud.ntnu.no>