docs(README.md): updated root readme for release 3
I updated the quickstart and running instructions to include the running of both the backend (server) and frontend (fxui). I also added images of JaCoCo test coverage for the already completed tests in core, shared and storage and of the apps new features and updated user interfaces. The previously used images from release 1 and 2 are added to folders inside the images folder.

#105