tests(fxui): complete FlashcardLoginControllerTest.java and almost complete FlashcardMainControllerTest.
The FlashcardLoginControllerTest.java is all done, but you need to run the tests spesifically for that file to get results in jacoco, which is a bit strange. Pushing so other teammembers can take a look.

FlashcardMainControllerTest. is not done since it misses some coverage on the branches. Pushing it so other teammembers can take a look.